dev_requirements.txt
black==23.3.0
isort==5.12.0
django-stubs==4.2.0
django-stubs-ext==4.2.0
